♻ refactor: 컨트롤러 인터페이스 도입을 통한 API 명세 및 로직 분리
- Swagger 어노테이션을 인터페이스로 이동하여 구현체 가독성 개선
- 보안 헤더가 필요한 API에 SecurityRequirement 추가
