+++ services/inventory-service/src/main/java/com/paystream/inventory/photo/controller/PhotoController.java
@@ -1,35 +1,31 @@
 package com.paystream.inventory.photo.controller;
 
-import com.paystream.inventory.photo.service.StorageService;
-import java.io.File;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import java.io.IOException;
-import java.nio.file.Files;
-import lombok.RequiredArgsConstructor;
-import org.springframework.core.io.FileSystemResource;
 import org.springframework.core.io.Resource;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
 
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/uploads")
-public class PhotoController {
+@Tag(name = "Photo", description = "이미지 처리 관련 API")
+public interface PhotoController {
 
-    private final StorageService storageService;
-
-    @GetMapping("/{imageName}")
-    public ResponseEntity<Resource> getImage(@PathVariable String imageName) throws IOException {
-        File file = storageService.getImageFile(imageName);
-        Resource resource = new FileSystemResource(file);
-
-        // 파일의 실제 MINE 타입을 자동으로 감지 (image/png, image/jpeg 등)
-        String contentType = Files.probeContentType(file.toPath());
-
-        return ResponseEntity.ok()
-                .contentType(
-                        MediaType.parseMediaType(
-                                contentType != null ? contentType : "application/octet-stream"))
-                .body(resource);
-    }
+    @Operation(
+            summary = "상품 이미지 조회",
+            description = "이미지 이름을 입력받아 해당 이미지 파일을 서버에서 스트리밍합니다.",
+            responses = {
+                @ApiResponse(
+                        responseCode = "200",
+                        description = "이미지 조회 성공",
+                        content = @Content(mediaType = "image/png") // 혹은 image/jpeg
+                        ),
+                @ApiResponse(
+                        responseCode = "404",
+                        description = "이미지를 찾을 수 없음",
+                        content = @Content // 에러 시에는 이미지 데이터가 없으므로 비워둠
+                        )
+            })
+    public ResponseEntity<Resource> getImage(@PathVariable String imageName) throws IOException;
 }
