+++ services/inventory-service/src/main/java/com/paystream/inventory/product/controller/ProductController.java
@@ -1,107 +1,66 @@
 package com.paystream.inventory.product.controller;
 
-import static com.paystream.inventory.utils.AuthUtils.getCurrentUserId;
-
 import com.paystream.core.BaseResponse;
 import com.paystream.inventory.product.dto.request.ProductCreateRequest;
 import com.paystream.inventory.product.dto.request.ProductDeleteRequest;
 import com.paystream.inventory.product.dto.request.ProductUpdateRequest;
 import com.paystream.inventory.product.dto.response.ProductDetailResponse;
 import com.paystream.inventory.product.dto.response.ProductResponse;
-import com.paystream.inventory.product.service.ProductCreateService;
-import com.paystream.inventory.product.service.ProductDeleteService;
-import com.paystream.inventory.product.service.ProductFindService;
-import com.paystream.inventory.product.service.ProductUpdateService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import java.time.LocalDate;
 import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.multipart.MultipartFile;
 
-@RestController
-@RequestMapping("/products")
-@RequiredArgsConstructor
-public class ProductController {
+@Tag(name = "Product", description = "상품 관련 API")
+public interface ProductController {
 
-    private final ProductFindService productFindService;
-    private final ProductCreateService productCreateService;
-    private final ProductUpdateService productUpdateService;
-    private final ProductDeleteService productDeleteService;
+    @Operation(summary = "상품 상세 조회", description = "상품 ID로 상세 정보를 조회합니다.")
+    BaseResponse<ProductDetailResponse> findById(@PathVariable Long productId);
 
-    @GetMapping("/{productId}")
-    public BaseResponse<ProductDetailResponse> findById(@PathVariable Long productId) {
-        ProductDetailResponse productDetail = productFindService.findProductDetail(productId);
-        return BaseResponse.ok(productDetail);
-    }
-
-    @GetMapping(
-            value = "/{productId}",
-            params = {"checkInDate", "checkOutDate"})
-    public BaseResponse<ProductDetailResponse> findByIdAndStock(
+    @Operation(summary = "상품 상세 및 재고 조회", description = "체크인/아웃 날짜를 포함하여 상품 상세 정보와 재고 상태를 조회합니다.")
+    BaseResponse<ProductDetailResponse> findByIdAndStock(
             @PathVariable Long productId,
             @RequestParam LocalDate checkInDate,
-            @RequestParam LocalDate checkOutDate) {
-        ProductDetailResponse productDetail =
-                productFindService.findProductWithDailyInventory(
-                        productId, checkInDate, checkOutDate);
-        return BaseResponse.ok(productDetail);
-    }
+            @RequestParam LocalDate checkOutDate);
 
     @Operation(
-            description = "상품 생성",
+            summary = "상품 생성",
+            description = "멀티파트 폼 데이터를 이용해 신규 상품을 생성합니다.",
             security = {@SecurityRequirement(name = "X-Auth-User-Id")})
-    @PostMapping(
-            consumes = MediaType.MULTIPART_FORM_DATA_VALUE,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public BaseResponse<Long> create(
+    BaseResponse<Long> create(
             HttpServletRequest request,
             @Valid @RequestPart("createRequest") ProductCreateRequest createRequest,
-            @RequestPart("file") List<MultipartFile> files) {
-        String hostId = getCurrentUserId(request);
-        Long savedProductId = productCreateService.create(hostId, createRequest, files);
+            @RequestPart("file") List<MultipartFile> files);
 
-        return BaseResponse.created(savedProductId);
-    }
-
-    @PutMapping("/{productId}")
-    public BaseResponse<Long> update(
+    @Operation(
+            summary = "상품 수정",
+            description = "기존 상품 정보를 업데이트합니다.",
+            security = {@SecurityRequirement(name = "X-Auth-User-Id")})
+    BaseResponse<Long> update(
             @PathVariable Long productId,
             HttpServletRequest request,
-            @Valid @RequestBody ProductUpdateRequest updateRequest) {
-        String hostId = getCurrentUserId(request);
-        Long updatedProductId = productUpdateService.update(hostId, productId, updateRequest);
+            @Valid @RequestBody ProductUpdateRequest updateRequest);
 
-        return BaseResponse.created(updatedProductId);
-    }
-
-    @DeleteMapping("/{productId}")
-    public BaseResponse<String> delete(
+    @Operation(
+            summary = "상품 삭제",
+            description = "상품을 삭제 처리합니다.",
+            security = {@SecurityRequirement(name = "X-Auth-User-Id")})
+    BaseResponse<String> delete(
             @PathVariable Long productId,
             HttpServletRequest request,
-            @Valid @RequestBody ProductDeleteRequest deleteRequest) {
-        String hostId = getCurrentUserId(request);
-        productDeleteService.delete(hostId, productId, deleteRequest);
+            @Valid @RequestBody ProductDeleteRequest deleteRequest);
 
-        return BaseResponse.ok("성공적으로 삭제되었습니다.");
-    }
-
-    /**
-     * 소유자가 갖고 있는 가게의 상품에 대한 정보만을 조회한다.
-     *
-     * @param request
-     * @return 기본적인 상품에 대한 정보
-     */
-    @GetMapping("/user/list")
-    public BaseResponse<List<ProductResponse>> findOwnedProductList(HttpServletRequest request) {
-        String hostId = getCurrentUserId(request);
-
-        List<ProductResponse> response = productFindService.listUserProducts(hostId);
-
-        return BaseResponse.ok(response);
-    }
+    @Operation(
+            summary = "내 상품 목록 조회",
+            description = "로그인한 사용자가 소유한 상품 목록을 조회합니다.",
+            security = {@SecurityRequirement(name = "X-Auth-User-Id")})
+    BaseResponse<List<ProductResponse>> findOwnedProductList(HttpServletRequest request);
 }
